Move build info from Board to App
