+++ src/App.css
@@ -92,3 +92,13 @@ input[type='search']::-webkit-search-cancel-button {
   /* icon size */
   background-size: 14px;
 }
+
+.debug {
+  color: silver;
+  font-size: xx-small;
+}
+
+.build-time {
+  z-index: 0;
+  text-align: right;
+}
